pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "This program allow to find differents between both files"
authors = ["nic11371 <nic11371@gmail.com>"]
license = "nic11371"
readme = "README.md"
packages = [
	{ include = "gendiff"},
]

[tool.poetry.dependencies]
python = "^3.10"
pytest-cov = "^5.0.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
